Introducing Base Custom App Sensor
